finish the unit teston ver1
