+++ tests/ftpserver.json
@@ -1,0 +1,6 @@
+{
+    "ftpserver": "nexycrypt.esy.es",
+    "username": "u431315912",
+    "password": "8dpv8LvjEE",
+    "comments": "The ftp server password is temporarily created and DO NOT use this value to do other things.The free web hosting will be closed or reset at the irregular time."
+}
